refactor(mllp-client): move the pure FIFO queue into util/ (with reconnect)
queue.ts is a pure, wire-agnostic deferred-promise FIFO with no I/O, no machine,
and it is currently unwired — the same "pure auxiliary, off the live path"
profile as util/reconnect.ts. Move it to src/util/queue.ts and refresh its stale
JSDoc (it described a drain driven by the now-deleted manager; reworded to the
generic "owner" and noted it is unwired).

src/ is now the live narrative — client, message, state, errors, index, node —
and src/util/ holds the pure helpers (queue, reconnect). 96 tests pass.
