+++ packages/mllp-client/src/util/queue.ts
@@ -1,34 +1,33 @@
 /**
- * `createSendQueue` — the FIFO send buffer for {@link MllpClient}.
+ * `createSendQueue` — a pure, wire-agnostic FIFO deferred-promise buffer.
  *
- * A pure, wire-agnostic deferred-promise queue. It holds the sends waiting
- * their turn on the wire and nothing else: it never touches the socket, the
+ * NOTE: currently UNWIRED. The client enforces single-flight directly (one send
+ * on the wire at a time; a concurrent send rejects `SEND_IN_PROGRESS`); this
+ * queue is kept for when FIFO concurrency is restored, at which point the
+ * client owns the drain loop and feeds it here.
+ *
+ * It buffers sends and nothing else: it never touches the socket, the
  * connection, an `AbortSignal`, or a timer. Each record carries only the wire
- * bytes, the correlation id, the ACK deadline (in ms — the manager turns it
- * into a real timer at the dispatch instant, not here), and the deferred
- * promise's settle handles.
+ * bytes, the correlation id, the ACK deadline (in ms — the owner turns it into
+ * a real timer at dispatch, not here), and the deferred promise's settle
+ * handles. `enqueue` returns the caller's real `Promise`, `take` hands the head
+ * record to the owner's drain loop, and `failAll` rejects everything still
+ * waiting (the owner supplies the error, e.g. `CLOSED`). An already-taken
+ * in-flight send is settled by the connection, not here.
  *
- * The manager owns all control flow — the single-flight drain loop, the
- * "is the wire up?" gate, building the per-send wire deadline, and calling
- * `connection.exchange`. The queue just buffers: `enqueue` returns the caller's
- * real `Promise`, `take` hands the head record to the drain loop, and `failAll`
- * rejects everything still waiting (the manager supplies the error, e.g.
- * `CLOSED`). The in-flight send — already `take`-n — is settled by the
- * connection, not here.
- *
- * Keeping the queue this dumb is what makes it unit-testable with no sockets,
- * no state machine, and no fakes: enqueue, then assert the returned promise
- * settles when the test calls the record's `resolve`/`reject`.
+ * Keeping it this dumb is what makes it unit-testable with no sockets, no state
+ * machine, and no fakes: enqueue, then assert the returned promise settles when
+ * the test calls the record's `resolve`/`reject`.
  *
  * @module
  */
 
-import type { MllpClientResponse } from "./message";
+import type { MllpClientResponse } from "../message";
 
 /**
  * One send waiting its turn on the wire. Carries the wire bytes, the
  * correlation id, the ACK deadline, and the deferred promise's settle handles —
- * no `AbortSignal`, no timer, no cleanup closures. The manager builds the
+ * no `AbortSignal`, no timer, no cleanup closures. The owner builds the
  * deadline when it dispatches the record and feeds it straight to
  * `connection.exchange`; the queue never sees it.
  */
@@ -36,7 +35,7 @@ export interface PendingSend {
   readonly framed: Uint8Array;
   readonly requestControlId: string;
   /**
-   * ACK-wait deadline (ms). The clock starts when the manager dispatches this
+   * ACK-wait deadline (ms). The clock starts when the owner dispatches this
    * record to the wire, not when it was enqueued.
    */
   readonly timeoutMs: number;
@@ -47,7 +46,7 @@ export interface PendingSend {
 export interface SendQueue {
   /**
    * Wrap a send in a real `Promise`, capture its settle handles on a buffered
-   * record, and return the promise. Does NOT drain — the manager kicks its loop
+   * record, and return the promise. Does NOT drain — the owner kicks its loop
    * after calling this. Pure bookkeeping: no signal, no timer, no wire.
    */
   enqueue(
@@ -57,14 +56,14 @@ export interface SendQueue {
   ): Promise<MllpClientResponse>;
   /**
    * Remove and return the head record (FIFO), or `undefined` if empty. The
-   * manager calls this in its drain loop to pull the next send for the wire.
-   * Once taken, a record is no longer the queue's responsibility — the manager
+   * owner calls this in its drain loop to pull the next send for the wire.
+   * Once taken, a record is no longer the queue's responsibility — the owner
    * (via the connection) settles it.
    */
   take(): PendingSend | undefined;
   /**
    * Reject every still-buffered record with `error` and empty the buffer. The
-   * manager supplies the error (`CLOSED` on drop/close); the queue only drains
+   * owner supplies the error (`CLOSED` on drop/close); the queue only drains
    * and rejects. The in-flight (already-taken) send is settled elsewhere.
    */
   failAll(error: Error): void;
@@ -86,7 +85,7 @@ export function createSendQueue(): SendQueue {
       timeoutMs: number
     ): Promise<MllpClientResponse> {
       // Deferred-send wrapper: the settle handles outlive this executor and
-      // fire from the manager's drain loop (or failAll), so a raw Promise is
+      // fire from the owner's drain loop (or failAll), so a raw Promise is
       // the right tool here.
       // oxlint-disable-next-line promise/avoid-new -- deferred-send wrapper
       return new Promise<MllpClientResponse>((resolve, reject) => {
